requirements.txt
openai~=1.37.0
requests~=2.32.3
transformers
torch
evalplus

python-dotenv~=1.0.1
h11~=0.14.0
pip~=24.1.2
attrs~=23.2.0
distro~=1.9.0
pillow~=10.4.0
filelock~=3.15.4
typing_extensions~=4.12.2
sniffio~=1.3.1
httpcore~=1.0.5
idna~=3.7
certifi~=2024.7.4
httpx~=0.27.0
pydantic~=2.8.2
anyio~=4.4.0
numpy~=1.26.4
pandas~=2.2.2
pydantic_core~=2.20.1